LPS-179592 Extract ScreenNavigationEntry methods from CommerceOrderItemVirtualSettingsScreenNavigationCategory to CommerceOrderItemVirtualSettingsScreenNavigationEntry and extends CommerceOrderItemVirtualSettingsScreenNavigationCategory in it
